preparation for new unique_award_key field
